♻️  Refactor: Refacor Season Components
