+++ src/lib/Season/Spring/index.tsx
@@ -1,7 +1,17 @@
-import React from "react";
+import React, { useEffect, useRef } from "react";
+import { Season as SeasonProps } from "../../types/common";
+import { falling } from "../utils/falling";
+import { EffectContainer } from "../styled";
 
-function Spring() {
-  return <div></div>;
-}
+const Spring: React.FC<SeasonProps> = ({ type = "auto" }) => {
+  const containerRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const container = containerRef.current;
+    falling(container);
+  }, []);
+
+  return <EffectContainer type={type} ref={containerRef} id="container" />;
+};
 
 export default Spring;
