+++ src/lib/Season/Summer/index.tsx
@@ -1,16 +1,16 @@
 import React from "react";
-import * as Styled from "./styled";
 import { Season as SeasonProps } from "../../types/common";
+import { SummerContainer } from "./styled";
 
 const Summer: React.FC<SeasonProps> = ({ children, type = "auto" }) => {
   return (
-    <Styled.Container type={type}>
+    <SummerContainer type={type}>
       <div className="ocean">
         <div className="wave"></div>
         <div className="wave"></div>
         <div>{children}</div>
       </div>
-    </Styled.Container>
+    </SummerContainer>
   );
 };
