+++ src/lib/Season/constants/index.tsx
@@ -1,11 +1,26 @@
-export const SPRING =
-  "linear-gradient(180deg, #91fd94 0%, #f5ffad 50.46%, #ffeaea 100%)";
+export const SPRING = "linear-gradient(180deg, #91fd94 0%, #f5ffad 50.46%, #ffeaea 100%)";
 
-export const SUMMER =
-  "radial-gradient(ellipse at center, rgba(255,254,234,1) 0%, rgba(255,254,234,1) 35%, #B7E8EB 100%)";
+export const SUMMER = "radial-gradient(ellipse at center, rgba(255,254,234,1) 0%, rgba(255,254,234,1) 35%, #B7E8EB 100%)";
 
-export const AUTUMN =
-  "linear-gradient(180deg, #FFD056 0%, #FD6670 99.99%, #FF001C 100%)";
+export const AUTUMN = "linear-gradient(180deg, #FFD056 0%, #FD6670 99.99%, #FF001C 100%)";
 
-export const WINTER =
-  "linear-gradient(180deg, #FFFFFF 0%, #FFEBF6 27.6%, #C1C6E9 100%)";
+export const WINTER = "linear-gradient(180deg, #FFFFFF 0%, #FFEBF6 27.6%, #C1C6E9 100%)";
+
+export const ASSET_ENDPOINT = "/asset/season";
+
+export const ASSET_SUMMER = {
+  wave: "https://s3-us-west-2.amazonaws.com/s.cdpn.io/85486/wave.svg",
+};
+
+export const ASSET_AUTUMN = {
+  item1: `${ASSET_ENDPOINT}/authum/leaf1.png`,
+  item2: `${ASSET_ENDPOINT}/authum/leaf2.png`,
+  item3: `${ASSET_ENDPOINT}/authum/leaf3.png`,
+  leaf4: `${ASSET_ENDPOINT}/authum/leaf4.png`,
+};
+
+export const ASSET_WINTER = {
+  item1: `${ASSET_ENDPOINT}/winter/snowflake1.png`,
+  item2: `${ASSET_ENDPOINT}/winter/snowflake2.png`,
+  item3: `${ASSET_ENDPOINT}/winter/snowflake3.png`,
+};
